Extract faceted perlin generator to core and made it work with existing world generator selection.
