+++ engine/src/main/java/org/terasology/world/chunks/internal/GeneratingChunkProvider.java
@@ -19,7 +19,7 @@ package org.terasology.world.chunks.internal;
 import org.terasology.math.Vector3i;
 import org.terasology.world.chunks.Chunk;
 import org.terasology.world.chunks.ChunkProvider;
-import org.terasology.world.generation.World;
+import org.terasology.world.generator.WorldGenerator;
 
 /**
  * Internal interface used within the chunk generation system, allows a chunk provider to manage "generation" (including
@@ -36,7 +36,7 @@ public interface GeneratingChunkProvider extends ChunkProvider {
      */
     void onChunkIsReady(Chunk chunk);
 
-    World getWorldGenerator();
+    WorldGenerator getWorldGenerator();
 
     Chunk getChunkUnready(Vector3i pos);
 }
