+++ engine/src/main/java/org/terasology/world/chunks/remoteChunkProvider/RemoteChunkProvider.java
@@ -38,8 +38,8 @@ import org.terasology.world.chunks.internal.GeneratingChunkProvider;
 import org.terasology.world.chunks.pipeline.AbstractChunkTask;
 import org.terasology.world.chunks.pipeline.ChunkGenerationPipeline;
 import org.terasology.world.chunks.pipeline.ChunkTask;
-import org.terasology.world.generation.World;
-import org.terasology.world.generation.WorldBuilder;
+import org.terasology.world.generator.WorldGenerator;
+import org.terasology.world.generator.internal.RemoteWorldGenerator;
 import org.terasology.world.internal.ChunkViewCore;
 import org.terasology.world.internal.ChunkViewCoreImpl;
 import org.terasology.world.propagation.light.InternalLightProcessor;
@@ -67,7 +67,7 @@ public class RemoteChunkProvider implements ChunkProvider, GeneratingChunkProvid
 
     private ChunkGenerationPipeline pipeline;
 
-    private World remoteWorld;
+    private WorldGenerator remoteWorld;
 
     private LightMerger<Chunk> lightMerger = new LightMerger<>(this);
 
@@ -75,7 +75,7 @@ public class RemoteChunkProvider implements ChunkProvider, GeneratingChunkProvid
         pipeline = new ChunkGenerationPipeline(new ChunkTaskRelevanceComparator());
         ChunkMonitor.fireChunkProviderInitialized(this);
 
-        remoteWorld = new WorldBuilder(0).build();
+        remoteWorld = new RemoteWorldGenerator();
     }
 
     public void subscribe(ChunkReadyListener chunkReadyListener) {
@@ -256,7 +256,7 @@ public class RemoteChunkProvider implements ChunkProvider, GeneratingChunkProvid
     }
 
     @Override
-    public World getWorldGenerator() {
+    public WorldGenerator getWorldGenerator() {
         //TODO: send this information over the wire
         return remoteWorld;
     }
