+++ engine/src/main/java/org/terasology/world/generation/GeneratingRegion.java
@@ -16,7 +16,6 @@
 package org.terasology.world.generation;
 
 import org.terasology.math.Region3i;
-import org.terasology.math.Vector3i;
 
 /**
  * @author Immortius
