+++ engine/src/main/java/org/terasology/world/generation/facets/SurfaceHeightFacet.java
@@ -16,8 +16,6 @@
 package org.terasology.world.generation.facets;
 
 import org.terasology.math.Region3i;
-import org.terasology.math.Vector2i;
-import org.terasology.math.Vector3i;
 import org.terasology.world.generation.Border3D;
 import org.terasology.world.generation.facets.base.BaseFieldFacet2D;
