+++ engine/src/main/java/org/terasology/world/generation/facets/base/BaseFacet2D.java
@@ -17,7 +17,6 @@ package org.terasology.world.generation.facets.base;
 
 import org.terasology.math.Rect2i;
 import org.terasology.math.Region3i;
-import org.terasology.math.Vector3i;
 import org.terasology.world.generation.Border3D;
 import org.terasology.world.generation.WorldFacet2D;
