+++ engine/src/main/java/org/terasology/world/generation/facets/base/BaseFacet3D.java
@@ -15,11 +15,8 @@
  */
 package org.terasology.world.generation.facets.base;
 
-import org.terasology.math.Rect2i;
 import org.terasology.math.Region3i;
-import org.terasology.math.Vector3i;
 import org.terasology.world.generation.Border3D;
-import org.terasology.world.generation.WorldFacet2D;
 import org.terasology.world.generation.WorldFacet3D;
 
 /**
