+++ engine/src/main/java/org/terasology/world/generation/facets/base/BaseFieldFacet3D.java
@@ -16,9 +16,7 @@
 package org.terasology.world.generation.facets.base;
 
 import com.google.common.base.Preconditions;
-import org.terasology.math.Border;
 import org.terasology.math.Region3i;
-import org.terasology.math.Vector2i;
 import org.terasology.math.Vector3i;
 import org.terasology.world.generation.Border3D;
