+++ engine/src/main/java/org/terasology/world/generation/facets/base/BaseObjectFacet2D.java
@@ -18,7 +18,6 @@ package org.terasology.world.generation.facets.base;
 import com.google.common.base.Preconditions;
 import org.terasology.math.Region3i;
 import org.terasology.math.Vector2i;
-import org.terasology.math.Vector3i;
 import org.terasology.world.generation.Border3D;
 
 import java.lang.reflect.Array;
