+++ engine/src/main/java/org/terasology/world/generation/facets/base/ObjectFacet2D.java
@@ -16,7 +16,6 @@
 package org.terasology.world.generation.facets.base;
 
 import org.terasology.math.Vector2i;
-import org.terasology.world.generation.WorldFacet;
 import org.terasology.world.generation.WorldFacet2D;
 
 /**
