+++ modules/Core/src/main/java/org/terasology/core/world/generator/facets/DensityFacet.java
@@ -13,19 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.facets;
+package org.terasology.core.world.generator.facets;
 
 import org.terasology.math.Region3i;
-import org.terasology.math.Vector3i;
 import org.terasology.world.generation.Border3D;
-import org.terasology.world.generation.facets.base.BaseBooleanFieldFacet3D;
+import org.terasology.world.generation.facets.base.BaseFieldFacet3D;
 
 /**
  * @author Immortius
  */
-public class PlantFacet extends BaseBooleanFieldFacet3D {
+public class DensityFacet extends BaseFieldFacet3D {
 
-    public PlantFacet(Region3i targetRegion, Border3D border) {
+    public DensityFacet(Region3i targetRegion, Border3D border) {
         super(targetRegion, border);
     }
 }
