+++ modules/Core/src/main/java/org/terasology/core/world/generator/facets/SolidityFacet.java
@@ -13,18 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.facets;
+package org.terasology.core.world.generator.facets;
 
 import org.terasology.math.Region3i;
 import org.terasology.world.generation.Border3D;
-import org.terasology.world.generation.facets.base.BaseFieldFacet2D;
+import org.terasology.world.generation.facets.base.BaseBooleanFieldFacet3D;
 
 /**
  * @author Immortius
  */
-public class HumidityFacet extends BaseFieldFacet2D {
+public class SolidityFacet extends BaseBooleanFieldFacet3D {
 
-    public HumidityFacet(Region3i targetRegion, Border3D border) {
+    public SolidityFacet(Region3i targetRegion, Border3D border) {
         super(targetRegion, border);
     }
 }
