+++ modules/Core/src/main/java/org/terasology/core/world/generator/perlinFacetProviders/PerlinBiomeProvider.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.perlin;
+package org.terasology.core.world.generator.perlinFacetProviders;
 
 import org.terasology.math.Vector2i;
-import org.terasology.math.Vector3i;
 import org.terasology.world.WorldBiomeProvider;
 import org.terasology.world.generation.Border3D;
 import org.terasology.world.generation.Facet;
