+++ modules/Core/src/main/java/org/terasology/core/world/generator/perlinFacetProviders/PerlinHillsAndMountainsProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.perlin;
+package org.terasology.core.world.generator.perlinFacetProviders;
 
 import org.terasology.math.TeraMath;
 import org.terasology.math.Vector2i;
