+++ modules/Core/src/main/java/org/terasology/core/world/generator/perlinFacetProviders/PerlinRiverProvider.java
@@ -13,49 +13,43 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.perlin;
+package org.terasology.core.world.generator.perlinFacetProviders;
 
-import org.terasology.math.Rect2i;
-import org.terasology.math.Region3i;
 import org.terasology.math.TeraMath;
-import org.terasology.math.Vector2i;
-import org.terasology.math.Vector3i;
 import org.terasology.utilities.procedural.BrownianNoise3D;
 import org.terasology.utilities.procedural.Noise3DTo2DAdapter;
 import org.terasology.utilities.procedural.PerlinNoise;
 import org.terasology.utilities.procedural.SubSampledNoise2D;
-import org.terasology.world.generation.Border3D;
+import org.terasology.world.generation.Facet;
 import org.terasology.world.generation.FacetProvider;
 import org.terasology.world.generation.GeneratingRegion;
-import org.terasology.world.generation.Produces;
-import org.terasology.world.generation.facets.HumidityFacet;
+import org.terasology.world.generation.Updates;
+import org.terasology.world.generation.facets.SurfaceHeightFacet;
 
 import javax.vecmath.Vector2f;
 
 /**
  * @author Immortius
  */
-@Produces(HumidityFacet.class)
-public class PerlinHumidityProvider implements FacetProvider {
+@Updates(@Facet(SurfaceHeightFacet.class))
+public class PerlinRiverProvider implements FacetProvider {
     private static final int SAMPLE_RATE = 4;
 
-    private SubSampledNoise2D humidityNoise;
+    private SubSampledNoise2D riverNoise;
 
     @Override
     public void setSeed(long seed) {
-        humidityNoise = new SubSampledNoise2D(new Noise3DTo2DAdapter(new BrownianNoise3D(new PerlinNoise(seed + 6), 8)), new Vector2f(0.0005f, 0.0005f), SAMPLE_RATE);
+        riverNoise = new SubSampledNoise2D(new Noise3DTo2DAdapter(new BrownianNoise3D(new PerlinNoise(seed + 2), 8)), new Vector2f(0.0008f, 0.0008f), SAMPLE_RATE);
     }
 
     @Override
     public void process(GeneratingRegion region) {
-        Border3D border = region.getBorderForFacet(HumidityFacet.class);
-        HumidityFacet facet = new HumidityFacet(region.getRegion(), border);
+        SurfaceHeightFacet facet = region.getRegionFacet(SurfaceHeightFacet.class);
+        float[] noise = riverNoise.noise(facet.getWorldRegion());
 
-        float[] noise = humidityNoise.noise(facet.getWorldRegion());
+        float[] surfaceHeights = facet.getInternal();
         for (int i = 0; i < noise.length; ++i) {
-            noise[i] = TeraMath.clamp((noise[i] + 1f) * 0.5f);
+            surfaceHeights[i] *= TeraMath.clamp(7f * (TeraMath.sqrt(Math.abs(noise[i])) - 0.1f) + 0.25f);
         }
-        facet.set(noise);
-        region.setRegionFacet(HumidityFacet.class, facet);
     }
 }
