+++ modules/Core/src/main/java/org/terasology/core/world/generator/perlinFacetProviders/PerlinTemperatureProvider.java
@@ -13,47 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.perlin;
+package org.terasology.core.world.generator.perlinFacetProviders;
 
-import org.terasology.math.Rect2i;
+import org.terasology.math.TeraMath;
 import org.terasology.utilities.procedural.BrownianNoise3D;
 import org.terasology.utilities.procedural.Noise3DTo2DAdapter;
 import org.terasology.utilities.procedural.PerlinNoise;
 import org.terasology.utilities.procedural.SubSampledNoise2D;
-import org.terasology.world.generation.Border3D;
 import org.terasology.world.generation.FacetProvider;
 import org.terasology.world.generation.GeneratingRegion;
 import org.terasology.world.generation.Produces;
-import org.terasology.world.generation.facets.SurfaceHeightFacet;
+import org.terasology.world.generation.facets.SeaLevelTemperatureFacet;
 
 import javax.vecmath.Vector2f;
 
 /**
  * @author Immortius
  */
-@Produces(SurfaceHeightFacet.class)
-public class PerlinBaseSurfaceProvider implements FacetProvider {
+@Produces(SeaLevelTemperatureFacet.class)
+public class PerlinTemperatureProvider implements FacetProvider {
     private static final int SAMPLE_RATE = 4;
 
-    private SubSampledNoise2D surfaceNoise;
+    private SubSampledNoise2D temperatureNoise;
 
     @Override
     public void setSeed(long seed) {
-        surfaceNoise = new SubSampledNoise2D(new Noise3DTo2DAdapter(new BrownianNoise3D(new PerlinNoise(seed), 8)), new Vector2f(0.004f, 0.004f), SAMPLE_RATE);
+        temperatureNoise = new SubSampledNoise2D(new Noise3DTo2DAdapter(new BrownianNoise3D(new PerlinNoise(seed + 5), 8)), new Vector2f(0.0005f, 0.0005f), SAMPLE_RATE);
     }
 
     @Override
     public void process(GeneratingRegion region) {
-        Border3D border = region.getBorderForFacet(SurfaceHeightFacet.class);
-        SurfaceHeightFacet facet = new SurfaceHeightFacet(region.getRegion(), border);
-        Rect2i processRegion = facet.getWorldRegion();
-        float[] noise = surfaceNoise.noise(processRegion);
+        SeaLevelTemperatureFacet facet = new SeaLevelTemperatureFacet(region.getRegion(), region.getBorderForFacet(SeaLevelTemperatureFacet.class));
+        float[] noise = this.temperatureNoise.noise(facet.getWorldRegion());
 
         for (int i = 0; i < noise.length; ++i) {
-            noise[i] = 32f + 32f * ((noise[i] + 1f) / 2f);
+            noise[i] = TeraMath.clamp((noise[i] + 1f) * 0.5f);
         }
 
         facet.set(noise);
-        region.setRegionFacet(SurfaceHeightFacet.class, facet);
+        region.setRegionFacet(SeaLevelTemperatureFacet.class, facet);
     }
 }
