+++ modules/Core/src/main/java/org/terasology/core/world/generator/rasterizers/SolidRasterizer.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.generation.rasterizer;
+package org.terasology.core.world.generator.rasterizers;
 
+import org.terasology.core.world.generator.facets.DensityFacet;
 import org.terasology.math.TeraMath;
 import org.terasology.math.Vector2i;
 import org.terasology.math.Vector3i;
@@ -26,7 +27,6 @@ import org.terasology.world.chunks.CoreChunk;
 import org.terasology.world.generation.Region;
 import org.terasology.world.generation.WorldRasterizer;
 import org.terasology.world.generation.facets.BiomeFacet;
-import org.terasology.world.generation.facets.DensityFacet;
 import org.terasology.world.generation.facets.SurfaceHeightFacet;
 
 /**
